Feat: UserRepository 인터페이스 생성
DB연동을 위해 새로운 repository를 생성해야 합니다.
수정 보다는 확장이 더 나은 코드라는 생각을 가지고 기존의 UserRepository 클래스를 MemoryUserRepository로 변경했고 UserRepository라는 인터페이스를 만들어 새로운 레퍼지토리를 만들 계획입니다.
