+++ src/test/java/kr/codesqaud/cafe/MemoryUserRepositoryTest.java
@@ -1,7 +1,7 @@
 package kr.codesqaud.cafe;
 
 import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.repository.UserRepository;
+import kr.codesqaud.cafe.repository.MemoryUserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,16 +10,16 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.*;
 
 
-class UserRepositoryTest {
-    UserRepository userRepository = new UserRepository();
+class MemoryUserRepositoryTest {
+    MemoryUserRepository memoryUserRepository = new MemoryUserRepository();
 
     @Test
     @DisplayName("유저 저장소에 유저가 제대로 추가된다.")
     void saveTest() {
         User user = new User("nag", "name", "asdf@gmail.com", "1234");
 
-        userRepository.save(user);
-        List<User> users = userRepository.getAllUsers();
+        memoryUserRepository.save(user);
+        List<User> users = memoryUserRepository.getAllUsers();
 
         assertThat(users).contains(user);
     }
